pkg/cover: normalize coverage paths during collection
Move the path normalization logic from `pkg/cover/backend/dwarf.go` into
a new `pkg/cover/backend/path.go` file to decouple it from DWARF parsing
and allow for independent reuse. Apply this normalization immediately
when coverage is collected in `pkg/aflow/action/crash/reproduce.go`
instead of deferring it to the syzlang tool. This ensures that coverage
data uses consistent, relative paths throughout the system from the
moment it is gathered, avoiding the need for downstream consumers to
handle environment-specific absolute paths, preventing LLM being
confused by path hiccups
