+++ pkg/cover/backend/dwarf.go
@@ -17,7 +17,6 @@ import (
 	"runtime"
 	"slices"
 	"strconv"
-	"strings"
 
 	"github.com/google/syzkaller/pkg/log"
 	"github.com/google/syzkaller/pkg/mgrconfig"
@@ -641,66 +640,6 @@ func readCoverPoints(target *targets.Target, info *symbolInfo, data []byte) ([2]
 		}
 	}
 	return pcs, nil
-}
-
-// Source files for Android may be split between two subdirectories: the common AOSP kernel
-// and the device-specific drivers: https://source.android.com/docs/setup/build/building-pixel-kernels.
-// Android build system references these subdirectories in various ways, which often results in
-// paths to non-existent files being recorded in the debug info.
-//
-// cleanPathAndroid() assumes that the subdirectories reside in `srcDir`, with their names being listed in
-// `delimiters`.
-// If one of the `delimiters` occurs in the `path`, it is stripped together with the path prefix, and the
-// remaining file path is appended to `srcDir + delimiter`.
-// If none of the `delimiters` occur in the `path`, `path` is treated as a relative path that needs to be
-// looked up in `srcDir + delimiters[i]`.
-func cleanPathAndroid(path, srcDir string, delimiters []string, existFn func(string) bool) (string, string) {
-	if len(delimiters) == 0 {
-		return "", ""
-	}
-	reStr := "(" + strings.Join(delimiters, "|") + ")(.*)"
-	re := regexp.MustCompile(reStr)
-	match := re.FindStringSubmatch(path)
-	if match != nil {
-		delimiter := match[1]
-		filename := match[2]
-		path := filepath.Clean(srcDir + delimiter + filename)
-		return filename, path
-	}
-	// None of the delimiters found in `path`: it is probably a relative path to the source file.
-	// Try to look it up in every subdirectory of srcDir.
-	for _, delimiter := range delimiters {
-		absPath := filepath.Clean(srcDir + delimiter + path)
-		if existFn(absPath) {
-			return path, absPath
-		}
-	}
-	return "", ""
-}
-
-func CleanPath(path string, kernelDirs *mgrconfig.KernelDirs, splitBuildDelimiters []string) (string, string) {
-	filename := ""
-
-	path = filepath.Clean(path)
-	aname, apath := cleanPathAndroid(path, kernelDirs.Src, splitBuildDelimiters, osutil.IsExist)
-	if aname != "" {
-		return aname, apath
-	}
-	absPath := osutil.Abs(path)
-	switch {
-	case strings.HasPrefix(absPath, kernelDirs.Obj):
-		// Assume the file was built there.
-		path = strings.TrimPrefix(absPath, kernelDirs.Obj)
-		filename = filepath.Join(kernelDirs.Obj, path)
-	case strings.HasPrefix(absPath, kernelDirs.BuildSrc):
-		// Assume the file was moved from buildDir to srcDir.
-		path = strings.TrimPrefix(absPath, kernelDirs.BuildSrc)
-		filename = filepath.Join(kernelDirs.Src, path)
-	default:
-		// Assume this is relative path.
-		filename = filepath.Join(kernelDirs.Src, path)
-	}
-	return strings.TrimLeft(filepath.Clean(path), "/\\"), filename
 }
 
 // objdump is an old, slow way of finding coverage points.
