+++ pkg/cover/backend/dwarf_test.go
@@ -37,70 +37,6 @@ func TestIsKcovBrokenInCompiler(t *testing.T) {
 	}
 }
 
-type CleanPathAndroidTest struct {
-	SrcDir     string
-	Delimiters []string
-	// Each test case is a triple of {path, epath, ename}, where path is passed to cleanPathAndroid(),
-	// and (epath, ename) is its expected return value.
-	Files    [][3]string
-	FnExists func(string) bool
-}
-
-func TestCleanPathAndroid(t *testing.T) {
-	tests := []CleanPathAndroidTest{
-		// Test that paths with "/aosp/" and "/private/" in them are normalized.
-		{
-			SrcDir:     "/src/kernel",
-			Delimiters: []string{"/aosp/", "/private/"},
-			Files: [][3]string{
-				{"/src/kernel/aosp/mm/mmap.c", "mm/mmap.c", "/src/kernel/aosp/mm/mmap.c"},
-				{"/src/kernel/out/cache/feedface/aosp/mm/mmap.c", "mm/mmap.c", "/src/kernel/aosp/mm/mmap.c"},
-				{"/src/kernel/out/cache/cafebabe/private/google_modules/module/mod.c", "google_modules/module/mod.c",
-					"/src/kernel/private/google_modules/module/mod.c"},
-				{"/some/other/path/aosp/mm/mmap.c", "mm/mmap.c", "/src/kernel/aosp/mm/mmap.c"},
-			},
-			FnExists: func(string) bool { return true },
-		},
-		// Test that for empty delimiters empty results are returned.
-		{
-			SrcDir:     "/src/kernel/",
-			Delimiters: []string{},
-			Files: [][3]string{
-				{"/src/kernel/mm/mmap.c", "", ""},
-			},
-			FnExists: func(string) bool { return true },
-		},
-		// Test that for path that does not contain a delimiter the result is constructed based on FnExists().
-		{
-			SrcDir:     "/src/kernel/",
-			Delimiters: []string{"/aosp/", "/private/"},
-			Files: [][3]string{
-				{"mm/mmap.c", "mm/mmap.c", "/src/kernel/aosp/mm/mmap.c"},
-			},
-			FnExists: func(s string) bool { return s == "/src/kernel/aosp/mm/mmap.c" },
-		},
-		// Test that for path that does not contain a delimiter the result is constructed based on FnExists().
-		{
-			SrcDir:     "/src/kernel/",
-			Delimiters: []string{"/aosp/", "/private/"},
-			Files: [][3]string{
-				{"mm/mmap.c", "mm/mmap.c", "/src/kernel/private/mm/mmap.c"},
-			},
-			FnExists: func(s string) bool { return s != "/src/kernel/aosp/mm/mmap.c" },
-		},
-	}
-	for _, test := range tests {
-		for _, files := range test.Files {
-			path, epath, ename := files[0], files[1], files[2]
-			rpath, rname := cleanPathAndroid(path, test.SrcDir, test.Delimiters, test.FnExists)
-			if (rpath != epath) || (rname != ename) {
-				t.Fatalf("cleanPathAndroid(`%s`, `%s`, %v, ...) unexpectedly returned (`%s`, `%s`) instead of (`%s`, `%s`)",
-					path, test.SrcDir, test.Delimiters, rpath, rname, epath, ename)
-			}
-		}
-	}
-}
-
 type NextCallTargetTest struct {
 	Arch      *Arch
 	Text      uint64
